backend/requirements.txt
sanic==23.12.1
tortoise-orm==0.21.6
aiomysql==0.2.0
openai==1.98.0
sanic-cors==2.2.0
pydantic==2.11.7
